Move TimeArbitraries to an arb package
