+++ modules/tests/src/test/scala/io/chrisdavenport/cats/time/instances/InstantTests.scala
@@ -4,9 +4,9 @@ package io.chrisdavenport.cats.time.instances
 import cats.tests.CatsSuite
 import cats.kernel.laws.discipline.HashTests
 import cats.kernel.laws.discipline.OrderTests
-import TimeArbitraries._
 import java.time.Instant
 import io.chrisdavenport.cats.time.instances.instant._
+import io.chrisdavenport.cats.time.arb.TimeArbitraries._
 
 class InstantTests extends CatsSuite {
   checkAll("Instant", HashTests[Instant].hash)
