+++ modules/tests/src/test/scala/io/chrisdavenport/cats/time/instances/LocalDateTests.scala
@@ -4,9 +4,9 @@ package io.chrisdavenport.cats.time.instances
 import cats.tests.CatsSuite
 import cats.kernel.laws.discipline.HashTests
 import cats.kernel.laws.discipline.OrderTests
-import TimeArbitraries._
 import java.time.LocalDate
 import io.chrisdavenport.cats.time.instances.localdate._
+import io.chrisdavenport.cats.time.arb.TimeArbitraries._
 
 class LocalDateTests extends CatsSuite {
   checkAll("LocalDate", HashTests[LocalDate].hash)
