+++ modules/tests/src/test/scala/io/chrisdavenport/cats/time/instances/LocalDateTimeTests.scala
@@ -4,9 +4,9 @@ package io.chrisdavenport.cats.time.instances
 import cats.tests.CatsSuite
 import cats.kernel.laws.discipline.HashTests
 import cats.kernel.laws.discipline.OrderTests
-import TimeArbitraries._
 import java.time.LocalDateTime
 import io.chrisdavenport.cats.time.instances.localdatetime._
+import io.chrisdavenport.cats.time.arb.TimeArbitraries._
 
 class LocalDateTimeTests extends CatsSuite {
   checkAll("LocalDateTime", HashTests[LocalDateTime].hash)
