+++ modules/tests/src/test/scala/io/chrisdavenport/cats/time/instances/LocalTimeTests.scala
@@ -4,9 +4,9 @@ package io.chrisdavenport.cats.time.instances
 import cats.tests.CatsSuite
 import cats.kernel.laws.discipline.HashTests
 import cats.kernel.laws.discipline.OrderTests
-import TimeArbitraries._
 import java.time.LocalTime
 import io.chrisdavenport.cats.time.instances.localtime._
+import io.chrisdavenport.cats.time.arb.TimeArbitraries._
 
 class LocalTimeTests extends CatsSuite {
   checkAll("LocalTime", HashTests[LocalTime].hash)
