+++ modules/tests/src/test/scala/io/chrisdavenport/cats/time/instances/MonthDayTests.scala
@@ -5,7 +5,7 @@ import java.time.MonthDay
 import cats.kernel.laws.discipline.{ HashTests, OrderTests }
 import cats.tests.CatsSuite
 import io.chrisdavenport.cats.time.instances.monthday._
-import TimeArbitraries._
+import io.chrisdavenport.cats.time.arb.TimeArbitraries._
 
 class MonthDayTests extends CatsSuite {
   checkAll("MonthDay", HashTests[MonthDay].hash)
