+++ modules/tests/src/test/scala/io/chrisdavenport/cats/time/instances/MonthTests.scala
@@ -4,7 +4,7 @@ import java.time.Month
 
 import cats.kernel.laws.discipline.{ HashTests, OrderTests }
 import cats.tests.CatsSuite
-import io.chrisdavenport.cats.time.instances.TimeArbitraries._
+import io.chrisdavenport.cats.time.arb.TimeArbitraries._
 import io.chrisdavenport.cats.time.instances.month._
 
 class MonthTests extends CatsSuite {
