+++ modules/tests/src/test/scala/io/chrisdavenport/cats/time/instances/YearMonthTests.scala
@@ -4,9 +4,9 @@ package io.chrisdavenport.cats.time.instances
 import cats.tests.CatsSuite
 import cats.kernel.laws.discipline.HashTests
 import cats.kernel.laws.discipline.OrderTests
-import TimeArbitraries._
 import java.time.YearMonth
 import io.chrisdavenport.cats.time.instances.yearmonth._
+import io.chrisdavenport.cats.time.arb.TimeArbitraries._
 
 class YearMonthTests extends CatsSuite {
   checkAll("YearMonth", HashTests[YearMonth].hash)
