+++ modules/tests/src/test/scala/io/chrisdavenport/cats/time/instances/YearTests.scala
@@ -3,9 +3,9 @@ package io.chrisdavenport.cats.time.instances
 import cats.tests.CatsSuite
 import cats.kernel.laws.discipline.HashTests
 import cats.kernel.laws.discipline.OrderTests
-import TimeArbitraries._
 import java.time.Year
 import io.chrisdavenport.cats.time.instances.year._
+import io.chrisdavenport.cats.time.arb.TimeArbitraries._
 
 class YearTests extends CatsSuite {
   checkAll("Year", HashTests[Year].hash)
