+++ modules/tests/src/test/scala/io/chrisdavenport/cats/time/instances/ZoneOffsetTests.scala
@@ -4,9 +4,9 @@ package io.chrisdavenport.cats.time.instances
 import cats.tests.CatsSuite
 import cats.kernel.laws.discipline.HashTests
 import cats.kernel.laws.discipline.OrderTests
-import TimeArbitraries._
 import java.time.ZoneOffset
 import io.chrisdavenport.cats.time.instances.zoneoffset._
+import io.chrisdavenport.cats.time.arb.TimeArbitraries._
 
 class ZoneOffsetTests extends CatsSuite {
   checkAll("ZoneOffset", HashTests[ZoneOffset].hash)
