+++ modules/tests/src/test/scala/io/chrisdavenport/cats/time/instances/ZonedDateTimeTests.scala
@@ -4,9 +4,9 @@ package io.chrisdavenport.cats.time.instances
 import cats.tests.CatsSuite
 import cats.kernel.laws.discipline.HashTests
 import cats.kernel.laws.discipline.OrderTests
-import TimeArbitraries._
 import java.time.ZonedDateTime
 import io.chrisdavenport.cats.time.instances.zoneddatetime._
+import io.chrisdavenport.cats.time.arb.TimeArbitraries._
 
 class ZonedDateTimeTests extends CatsSuite {
   checkAll("ZonedDateTime", HashTests[ZonedDateTime].hash)
